reuse existing functions from lib rather than writing new versions.
