+++ test/unit/three.addons.unit.js
@@ -40,4 +40,5 @@ import './addons/tsl/TSLColorAdjustmentExtra.tests.js';
 import './addons/tsl/TSLToneMapping.tests.js';
 import './addons/tsl/TSLProceduralUtils.tests.js';
 import './addons/tsl/TSLBSDFLightingRemainder.tests.js';
+import './addons/tsl/TSL.Irradiance.tests.js';
 import './addons/tsl/TSLUtilsMisc.tests.js';
